Deleting files in temporary folder upon logout

+++ f8d890ce88bd1791b6eaddf06e58ceb5/logout.php
@@ -12,8 +12,9 @@
     session_destroy();
     // Remove all temporary files
     $tmpFiles = glob("tmp/*");
-    foreach($f as $tmpFiles){
-		//if(is_file($f))
+    echo count($tmpFiles);
+    foreach($tmpFiles as $f){
+		if(is_file($f))
 			unlink($f);
 	}
     header("Location: ../notify.php?mode=signout");
